// client/src/models/IVideoCard.ts
export interface IVideoCard {
    id: number;
    name: string;
    brand: string;
    model: string;
    memory_volume: number;
    tire_capacity: number;
    memory_type: string;
    power_consumption: number;
    memory_frequency: number;
    vga: number;
    hdmi: number;
    dvi: number;
    display_port: number;
    video_card_length: number;
    price: number;
    img: string;
}